+++ backend/src/main/java/ca/metricalsky/winston/exception/ProblemError.java
@@ -1,0 +1,12 @@
+package ca.metricalsky.winston.exception;
+
+public record ProblemError(
+        String type,
+        String detail,
+        Location location
+) {
+
+    public ProblemError(String type, String detail) {
+        this(type, detail, null);
+    }
+}